WIP: ml-190 - tensorflow recal
Summary:
alright, don't kill me for the size of the diff. this is a big architecture implementation so after this diffs should be much more manageble. putting it up so people can start looking through.
currently still in progress, but majority of work is done. specifically overall architecture and pruning ops for tensorflow. things to still be tackled: logging and test cases

Test Plan: unit / integration tests to be added

Reviewers: tuan, mgoin, alexm, ben

Reviewed By: mgoin

Subscribers: #core, #python_ml

Differential Revision: https://phabricator.neuralmagic.com:8021/D1620