WIP - ML207: Loss sensitivity analysis for Tensorflow
Summary:
Push this up now to collaborate with Mark on a fix for the pruning ops

The unit test passes now but the loss does not change with changing sparsity. I looked into the masks and saw that masks coming out of update ops are not correct.
I'll add more tests after we fix this issue; also will add GPU support for the code.

Test Plan: Preliminary test.

Reviewers: mark.kurtz, mgoin, alexm, ben

Reviewed By: mark.kurtz

Subscribers: #core

Differential Revision: https://phabricator.neuralmagic.com:8021/D1681