docs refactoring for remaining modules
Summary:
a lot of line changes, but mainly only in the documentation of it.
Note, I moved the recal packages to a flat structure to appear better in the docs (so much nesting caused issueus).
but after one more big diff ffor test cases, we're done with the big diffs!

Reviewers: mgoin, alexm, ben, tuan

Reviewed By: mgoin, tuan

Subscribers: #core, #python_ml

Differential Revision: https://phabricator.neuralmagic.com:8021/D1688