Learning rate modifiers for TensorFlow
Summary:
This revision is to support learning rate modifiers in Tensorflow. As with PyTorch version, it supports SetLearningRateModifier to set a fixed learning rate starting at a given epoch, and LearningRateModifier that allows the usage of built-in schedulers in Tensorflow (ExponentialDecay, InverseTimeDecay etc).

Modifiers can be defined in sequence, in which case if two or more are active at a global time step then the learning rate will be set based on the last modifier. As an example, the following sequence (as part of the tests),

[(setlr_01_spec, 0, -1, "SetLearningRateModifier"), (setlr_02_spec, 1, -1, "SetLearningRateModifier"), (expo_spec, 1, 3, "LearningRateModifier"), (inv_spec, 3, -1, "LearningRateModifier")],

defines 4 modifiers such that the second and third ones overlaps starting at epoch 1, and therefore the third one (expo_spec) takes over.

Known limitations: The following needs to be added: (1) Logging to Tensorboard, (2) Optional init LR for LearningRateModifier to be compatible with PyTorch version (which on the other hand can be set using a preceding SetLearningRateModifier)

Fix Lint issue

Test Plan: Added unit tests for individual and sequence of modifiers

Reviewers: mark.kurtz, alexm, mgoin, ben

Reviewed By: mark.kurtz

Subscribers: #core

Differential Revision: https://phabricator.neuralmagic.com:8021/D1740