Remove tensorflow-onnx submodule
Summary: The only change necessary for neuralmagic-integration to work is the new name for tf2onnx loader used in the exporter, but this essentially means the current fork of tensorflow-onnx submodule needs to be upgraded. As we discussed, this change removes the submodule and install tf2onnx as part of the requirements.

Test Plan: Ran unit tests, all passed.

Reviewers: mark.kurtz, alexm, mgoin, ben

Reviewed By: mark.kurtz

Subscribers: #core

Differential Revision: https://phabricator.neuralmagic.com:8021/D2033