Support for finding trivially nested Conv/Gemm weights in onnx utils
Summary:
Bug found by @kevinaer where a TF exported mobilenet v3 model would crash in the model analyzer because it's weights came from a series of reshapes/transposes within the graph.

This diff adds support to search for a model's weights as long as it comes from a series of only reshapes/transposes.  In this case, the given weights will be the original model weight, not the transposed/reshaped version.

Test Plan:
 - manually ran tests/onnx
 - model analyzer now runs successfully for the tf exported mobilenet v3 model

Reviewers: mark.kurtz, kevinaer, tuan, mgoin, alexm, dhuang

Reviewed By: mark.kurtz, kevinaer

Subscribers: kevinaer, #core

Differential Revision: https://phabricator.neuralmagic.com:8021/D2652