Implementation of Top-KAST pruning algorithm (#1674)
Implementation of Top-KAST (Top-K always sparse training) pruning algorithm, published in https://arxiv.org/abs/2106.03517

The idea of the algorithm is that weights (forward step) and gradients (backward step) are pruned at two different sparsity levels (forward is sparser), allowing some weights to grow even as they are not used in the forward pass. Additionally, both forward-enabled and backward-enabled weights are decayed by the L2 norm at every step, allowing even weights that are not otherwise updated to potentially grow relative to those that are. The forward and backward masks are updated at each step in the original paper, but every ~100 steps is better for a batch size of 400.

Testing: added unit tests. In addition, ran a 100-epoch 90 forward/80 backward sparsity run, obtaining a 72.5% top-1 accuracy model. The paper reports 73.0% top-1 accuracy, but uses an impractically large batch size. Performance-wise, recomputing the mask at every step leads to a ~15% slowdown when training with a batch size of 400 on two Tesla V1000s, which seems tolerable (especially because in practice we do it less often).

---------

Co-authored-by: Alexandre Marques <alexandre@neuralmagic.com>