Implementation of powerpropagation as a modifier (#1685)
Implement Powerpropagation (raising weights to a power in forward pass), see https://arxiv.org/abs/2110.00296

The functionality is implemented as a Modifier that, when activated, wraps pruned modules (for now, only Conv2d or Linear, but can easily be extended later) in a PowerpropagationWrapper, which does the exponentiation. Once the modifier is deactivated, the architecture change is undone.

Testing: wrote new tests, including for the architecture change and corresponding weight updates. Ran a few epochs on imagenet to make sure that training was looking OK (loss going down/accuracy going up).